chore(wmg): refactor filters top component
